fix bc break in RequestSignatureListener
