+++ Security/Firewall/RequestSignatureListener.php
@@ -2,71 +2,28 @@
 
 namespace Rezzza\SecurityBundle\Security\Firewall;
 
-use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationException;
-use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
-use Symfony\Component\Security\Core\Authentication\Token\AnonymousToken;
-use Rezzza\SecurityBundle\Security\RequestSignatureToken;
-use Psr\Log\LoggerInterface;
 
-class RequestSignatureListener
-{
-    protected $tokenStorage;
-    protected $authenticationManager;
-    protected $signatureQueryParameters;
-    protected $ignored;
-    protected $logger;
-
-    public function __construct(
-        TokenStorageInterface $tokenStorage,
-        AuthenticationManagerInterface $authenticationManager,
-        SignatureQueryParameters $signatureQueryParameters,
-        $ignored,
-        LoggerInterface $logger = null
-    )
+// Symfony < 4.3
+if (!class_exists(RequestEvent::class)) {
+    class RequestSignatureListener
     {
-        $this->tokenStorage = $tokenStorage;
-        $this->authenticationManager = $authenticationManager;
-        $this->signatureQueryParameters = $signatureQueryParameters;
-        $this->ignored = $ignored;
-        $this->logger = $logger;
+        use RequestSignatureListenerTrait;
+
+        public function handle(GetResponseEvent $event)
+        {
+            $this->handleOrInvoke($event);
+        }
     }
-
-    public function __invoke(RequestEvent $event)
+} else {
+    class RequestSignatureListener
     {
-        if (true === $this->ignored) {
-            if (null !== $this->tokenStorage->getToken()) {
-                return;
-            }
+        use RequestSignatureListenerTrait;
 
-            $this->tokenStorage->setToken(new AnonymousToken('request_signature', 'anon.', array()));
-
-            return;
+        public function __invoke(RequestEvent $event)
+        {
+            $this->handleOrInvoke($event);
         }
-
-        $request = $event->getRequest();
-        $authToken = new RequestSignatureToken;
-        $authToken->signature = $request->get($this->signatureQueryParameters->getNonceQueryParameter());
-        $authToken->signatureTime = $request->get($this->signatureQueryParameters->getTimeQueryParameter());
-        $authToken->requestMethod = $request->server->get('REQUEST_METHOD');
-        $authToken->requestHost = $request->server->get('HTTP_HOST');
-        $authToken->requestPathInfo = $request->getPathInfo();
-        $authToken->requestContent = rawurldecode($request->getContent());
-
-        try {
-            return $this->tokenStorage->setToken(
-                $this->authenticationManager->authenticate($authToken)
-            );
-        } catch (AuthenticationException $e) {
-            if ($this->logger) {
-                $this->logger->info(sprintf('Authentication request failed: %s', $e->getMessage()));
-            }
-        }
-
-        $response = new Response();
-        $response->setStatusCode(403);
-        $event->setResponse($response);
     }
 }
